requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
mysqlclient==2.1.0
python-dotenv==0.19.0
click==8.0.3
python-dateutil==2.8.2